+++ src/BigCommerce/Api/Orders/RefundsApi.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace BigCommerce\ApiV3\Orders;
+namespace BigCommerce\ApiV3\Api\Orders;
 
 use BigCommerce\ApiV3\Api\Generic\V3ApiBase;
 use BigCommerce\ApiV3\ResourceModels\Order\OrderRefundItem;
