+++ src/BigCommerce/Client.php
@@ -2,7 +2,7 @@
 
 namespace BigCommerce\ApiV3;
 
-use BigCommerce\ApiV3\Orders\OrdersApi;
+use BigCommerce\ApiV3\Api\Orders\OrdersApi;
 use BigCommerce\ApiV3\Api\Customers\CustomersApi;
 use BigCommerce\ApiV3\Payments\PaymentsProcessingApi;
 use BigCommerce\ApiV3\PriceLists\PriceListsApi;
